fix(review): resolve remaining acknowledged items (#147)
Closes every item that was previously in the residuals file's
'Acknowledged (no code change)' bucket. No deferrals remain.

H3: warnDriftOnce throttles missing-agentId / missing-campaignId
warnings in use-events.ts to one per scope+event-type per 60s
cooldown. The first warn per drift signature still surfaces the
problem; a misbehaving backend can no longer flood the console.

S3: CampaignDagView height is now content-aware (estimateHeight
between MIN_DAG_HEIGHT and MAX_DAG_HEIGHT) and panOnDrag={true} so
operators can reach nodes outside the initial fitView. Node dragging
stays disabled — pan is a viewport-only operation in a read-only
graph.

S4: getCampaignTaskStats, listActiveAgentsByCampaign, deleteCampaign,
and the DeleteCampaignResult type moved into the new
services/campaign-dashboard.ts. services/campaigns.ts is back under
the 800-line guideline (686 lines from 892). Re-exports preserve the
existing import path for the route handler and tests.

Q4: The isolated-phase skip stub in dashboard-campaigns-routes.test.ts
no longer passes silently when run outside its phase. It now
console.warn's and asserts on the env gate so a CI misconfiguration
cannot drop the route coverage while reading green.

Residuals file updated to reflect that nothing is deferred.
